# docs/changelog/v1/openapi/agros-api-v1.0.yaml
openapi: 3.1.0
info:
  title: Agri OS Core API
  summary: Deterministic core API for Agri OS
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      tags:
      - ProjectScopes
      summary: List Project Scopes
      operationId: list_project_scopes_api_v1_projects_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - ProjectScopes
      summary: Create Project Scope
      operationId: create_project_scope_api_v1_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectScopeRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}:
    get:
      tags:
      - ProjectScopes
      summary: Get Project Scope
      operationId: get_project_scope_api_v1_projects__project_scope_id__get
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeDetail'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/projects/{project_scope_id}/activate:
    post:
      tags:
      - ProjectScopes
      summary: Activate Project Scope
      operationId: activate_project_scope_api_v1_projects__project_scope_id__activate_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateProjectScopeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/pause:
    post:
      tags:
      - ProjectScopes
      summary: Pause Project Scope
      operationId: pause_project_scope_api_v1_projects__project_scope_id__pause_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PauseProjectScopeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/close:
    post:
      tags:
      - ProjectScopes
      summary: Close Project Scope
      operationId: close_project_scope_api_v1_projects__project_scope_id__close_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseProjectScopeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/archive:
    post:
      tags:
      - ProjectScopes
      summary: Archive Project Scope
      operationId: archive_project_scope_api_v1_projects__project_scope_id__archive_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveProjectScopeRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectScopeResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/assignments:
    post:
      tags:
      - ProjectScopes
      summary: Create Project Assignment
      operationId: create_project_assignment_api_v1_projects__project_scope_id__assignments_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectAssignmentRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAssignmentResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - ProjectScopes
      summary: List Project Assignments
      operationId: list_project_assignments_api_v1_projects__project_scope_id__assignments_get
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAssignmentListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/assignments/{project_assignment_id}/end:
    post:
      tags:
      - ProjectScopes
      summary: End Project Assignment
      operationId: end_project_assignment_api_v1_projects__project_scope_id__assignments__project_assignment_id__end_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      - name: project_assignment_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Assignment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndProjectAssignmentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAssignmentResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/contributions:
    post:
      tags:
      - ProjectScopes
      summary: Record Project Contribution
      operationId: record_project_contribution_api_v1_projects__project_scope_id__contributions_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordProjectContributionRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - ProjectScopes
      summary: List Project Contributions
      operationId: list_project_contributions_api_v1_projects__project_scope_id__contributions_get
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/contributions/{project_contribution_event_id}/confirm:
    post:
      tags:
      - ProjectScopes
      summary: Confirm Project Contribution
      operationId: confirm_project_contribution_api_v1_projects__project_scope_id__contributions__project_contribution_event_id__confirm_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      - name: project_contribution_event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Contribution Event Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmProjectContributionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/contributions/{project_contribution_event_id}/reject:
    post:
      tags:
      - ProjectScopes
      summary: Reject Project Contribution
      operationId: reject_project_contribution_api_v1_projects__project_scope_id__contributions__project_contribution_event_id__reject_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      - name: project_contribution_event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Contribution Event Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectProjectContributionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContributionResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_scope_id}/cost-records:
    post:
      tags:
      - ProjectScopes
      summary: Create Project Cost Record
      operationId: create_project_cost_record_api_v1_projects__project_scope_id__cost_records_post
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectCostRecordRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCostRecordResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - ProjectScopes
      summary: List Project Cost Records
      operationId: list_project_cost_records_api_v1_projects__project_scope_id__cost_records_get
      parameters:
      - name: project_scope_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Scope Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectCostRecordListResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Aggregate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateProjectScopeRequest:
      properties:
        organizationId:
          type: string
          title: Organizationid
        name:
          type: string
          title: Name
        projectScopeType:
          $ref: '#/components/schemas/ProjectScopeType'
        seasonYear:
          anyOf:
          - type: string
          - type: 'null'
          title: Seasonyear
        ownerActorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneractorid
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        parentProjectScopeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentprojectscopeid
        metadata:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Metadata
        meta:
          anyOf:
          - $ref: '#/components/schemas/Meta'
          - type: 'null'
      type: object
      required:
      - organizationId
      - name
      - projectScopeType
      title: CreateProjectScopeRequest
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    ProjectScopeDetail:
      properties:
        projectScopeId:
          type: string
          title: Projectscopeid
        organizationId:
          type: string
          title: Organizationid
        projectScopeCode:
          type: string
          title: Projectscopecode
        name:
          type: string
          title: Name
        projectScopeType:
          $ref: '#/components/schemas/ProjectScopeType'
        status:
          $ref: '#/components/schemas/ProjectScopeStatus'
        seasonYear:
          anyOf:
          - type: string
          - type: 'null'
          title: Seasonyear
        ownerActorId:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneractorid
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        parentProjectScopeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Parentprojectscopeid
        metadata:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Metadata
        updatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - projectScopeId
      - organizationId
      - projectScopeCode
      - name
      - projectScopeType
      - status
      title: ProjectScopeDetail
    ProjectScopeListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectScopeSummary'
          type: array
          title: Items
      type: object
      required:
      - items
      title: ProjectScopeListResponse
    ProjectScopeResponse:
      properties:
        data:
          $ref: '#/components/schemas/ProjectScopeDetail'
      type: object
      required:
      - data
      title: ProjectScopeResponse
...